Develop

+++ restaurant/repository/mongorepo.py
@@ -31,8 +31,9 @@ class MongoRepo:
     def get(self,dish_id):
         collection = self.db.dishes
         dish = collection.find_one({'id': dish_id})
-        result = self._create_dish_objects(dish)
-        return result
+        if dish:
+            result = self._create_dish_objects(dish)
+            return result
 
     def post(self,dish):
         self.db.dishes.insert_one(dish)
@@ -61,6 +62,6 @@ class MongoRepo:
             return self.list()
         
     def delete(self,dish_id):
-        self.db.dishes.delete_one({"id": dish_id})
-        return self.list()
-    
+        if self.db.dishes.find_one({"id": dish_id}):
+            self.db.dishes.delete_one({"id": dish_id})
+            return self.list()
